[core] Encapsulate split serialization protocols
