+++ paimon-core/src/main/java/org/apache/paimon/table/source/IncrementalSplit.java
@@ -23,6 +23,7 @@ import org.apache.paimon.io.DataFileMeta;
 import org.apache.paimon.io.DataFileMetaSerializer;
 import org.apache.paimon.io.DataInputView;
 import org.apache.paimon.io.DataInputViewStreamWrapper;
+import org.apache.paimon.io.DataOutputView;
 import org.apache.paimon.io.DataOutputViewStreamWrapper;
 import org.apache.paimon.utils.FunctionWithIOException;
 
@@ -191,7 +192,27 @@ public class IncrementalSplit implements Split {
     }
 
     private void writeObject(ObjectOutputStream objectOutputStream) throws IOException {
-        DataOutputViewStreamWrapper out = new DataOutputViewStreamWrapper(objectOutputStream);
+        serialize(new DataOutputViewStreamWrapper(objectOutputStream));
+    }
+
+    private void readObject(ObjectInputStream objectInputStream)
+            throws IOException, ClassNotFoundException {
+        assign(deserialize(new DataInputViewStreamWrapper(objectInputStream)));
+    }
+
+    protected void assign(IncrementalSplit other) {
+        snapshotId = other.snapshotId;
+        partition = other.partition;
+        bucket = other.bucket;
+        totalBuckets = other.totalBuckets;
+        beforeFiles = other.beforeFiles;
+        beforeDeletionFiles = other.beforeDeletionFiles;
+        afterFiles = other.afterFiles;
+        afterDeletionFiles = other.afterDeletionFiles;
+        isStreaming = other.isStreaming;
+    }
+
+    public void serialize(DataOutputView out) throws IOException {
         out.writeInt(VERSION);
         out.writeLong(snapshotId);
         serializeBinaryRow(partition, out);
@@ -216,39 +237,49 @@ public class IncrementalSplit implements Split {
         out.writeBoolean(isStreaming);
     }
 
-    private void readObject(ObjectInputStream objectInputStream)
-            throws IOException, ClassNotFoundException {
-        DataInputViewStreamWrapper in = new DataInputViewStreamWrapper(objectInputStream);
+    public static IncrementalSplit deserialize(DataInputView in) throws IOException {
         int version = in.readInt();
         if (version != VERSION) {
             throw new UnsupportedOperationException("Unsupported version: " + version);
         }
 
-        snapshotId = in.readLong();
-        partition = deserializeBinaryRow(in);
-        bucket = in.readInt();
-        totalBuckets = in.readInt();
+        long snapshotId = in.readLong();
+        BinaryRow partition = deserializeBinaryRow(in);
+        int bucket = in.readInt();
+        int totalBuckets = in.readInt();
 
         DataFileMetaSerializer dataFileMetaSerializer = new DataFileMetaSerializer();
         FunctionWithIOException<DataInputView, DeletionFile> deletionFileSerializer =
                 DeletionFile::deserialize;
 
         int beforeNumber = in.readInt();
-        beforeFiles = new ArrayList<>(beforeNumber);
+        List<DataFileMeta> beforeFiles = new ArrayList<>(beforeNumber);
         for (int i = 0; i < beforeNumber; i++) {
             beforeFiles.add(dataFileMetaSerializer.deserialize(in));
         }
 
-        beforeDeletionFiles = DeletionFile.deserializeList(in, deletionFileSerializer);
+        List<DeletionFile> beforeDeletionFiles =
+                DeletionFile.deserializeList(in, deletionFileSerializer);
 
         int fileNumber = in.readInt();
-        afterFiles = new ArrayList<>(fileNumber);
+        List<DataFileMeta> afterFiles = new ArrayList<>(fileNumber);
         for (int i = 0; i < fileNumber; i++) {
             afterFiles.add(dataFileMetaSerializer.deserialize(in));
         }
 
-        afterDeletionFiles = DeletionFile.deserializeList(in, deletionFileSerializer);
+        List<DeletionFile> afterDeletionFiles =
+                DeletionFile.deserializeList(in, deletionFileSerializer);
 
-        isStreaming = in.readBoolean();
+        boolean isStreaming = in.readBoolean();
+        return new IncrementalSplit(
+                snapshotId,
+                partition,
+                bucket,
+                totalBuckets,
+                beforeFiles,
+                beforeDeletionFiles,
+                afterFiles,
+                afterDeletionFiles,
+                isStreaming);
     }
 }
